Removed trackpad-pan controller
